// packages/main/src/models/formItems/enums/FormItemTypeEnum.ts
enum FormItemTypeEnum {
  CUSTOM_INPUT = 'CUSTOM_INPUT',
  CUSTOM_TEXTAREA = 'CUSTOM_TEXTAREA',
  CUSTOM_SINGLE_SELECT = 'CUSTOM_SINGLE_SELECT',
  CUSTOM_DIVIDER_LINE = 'CUSTOM_DIVIDER_LINE',
  CUSTOM_UPLOAD = 'CUSTOM_UPLOAD',
  CUSTOM_CHECKBOX = 'CUSTOM_CHECKBOX',
  CUSTOM_SWITCH = 'CUSTOM_SWITCH',
  CUSTOM_COLOR_SELECT = 'CUSTOM_COLOR_SELECT',
}

export default FormItemTypeEnum